packages/scripts/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
datasets==2.20.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langsmith==0.1.83
llama-cloud==0.0.6
llama-index==0.10.52
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.52.post1
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.27
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.5
marshmallow==3.21.3
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.35.10
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
pypdf==4.2.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ragas==0.1.10
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
striprtf==0.0.26
tenacity==8.4.2
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
